Refactor WOM traits and frame pointer support into dedicated module
- Created wom_traits.rs module containing:
  - Frame Bus System (FrameBus, FrameState, FrameBridge)
  - VmChipWrapperWom and associated traits
  - VmAdapterChipWom and VmCoreChipWom traits
  - AdapterRuntimeContextWom for frame pointer transitions
- Cleaned up extension.rs by removing duplicated code
- Updated imports and fixed compilation warnings
- Improved code organization for Womir-specific functionality
