+++ rv32im-wom/circuit/src/lib.rs
@@ -35,5 +35,8 @@ pub use shift::*;
 mod extension;
 pub use extension::*;
 
+mod wom_traits;
+pub use wom_traits::*;
+
 #[cfg(any(test, feature = "test-utils"))]
 mod test_utils;
